+++ toolbox/diver/find-contributors-in-git-logs.sh
@@ -7,7 +7,7 @@
 #
 # Author: Pierre-Yves LAPERSONNE <pierreyves(dot)lapersonne(at)orange(dot)com> et al.
 
-# Version.............: 1.2.0
+# Version.............: 1.2.1
 # Since...............: 11/05/2020
 # Description.........: Looks for words (defined in dedicated file) in git logs
 #
@@ -21,7 +21,7 @@
 #
 
 
-VERSION="1.2.0"
+VERSION="1.2.1"
 SCRIPT_NAME="find-contributors-in-git-logs"
 
 # -------------
@@ -232,6 +232,11 @@ cd "$git_based_project"
 git_log_file="../$GIT_LOG_TEMP_FILE"
 if [ -f $git_log_file ]; then
     rm $git_log_file
+fi
+
+if [ "$( git log --oneline -5 2>/dev/null | wc -l )" -eq 0 ]; then
+    echo "Warning: Project '$git_based_project' is a git repository without any commit, that's weird"
+    NormalExit
 fi
 
 git log --since=$git_log_limit > $git_log_file